backend/requirements.txt

fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
motor==3.1.2
pyjwt==2.7.0
